% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{approx_equal}
\alias{approx_equal}
\alias{\%~\%}
\title{Approximate equal}
\usage{
approx_equal(x, y, tol = .Machine$double.eps^0.5)

x \%~\% y
}
\arguments{
\item{x}{Numeric scalar.}

\item{y}{Numeric scalar.}

\item{tol}{Tolerance.}
}
\value{
Boolean
}
\description{
Compute whether x and y are approximately equal given a tolerance level
}
\examples{
approx_equal(1, 1)
1 \%~\% (1 + 1e-16)
1 \%~\% 1.01
}
