% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_colnames.R
\name{change_colnames}
\alias{change_colnames}
\title{Change column names of a dataframe}
\usage{
change_colnames(df, current_names, new_names)
}
\arguments{
\item{df}{dataframe}

\item{current_names}{vector of column names to change}

\item{new_names}{vector of new names}
}
\value{
dataframe with new column names
}
\description{
Change column names of a dataframe
}
