% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_post.R
\name{check_model_sensitivity}
\alias{check_model_sensitivity}
\title{Checking how posterior are influenced by the prior}
\usage{
check_model_sensitivity(prior, post, param)
}
\arguments{
\item{prior}{Dataframe of prior parameter estimates (with columns Variable, Index, Mean and sd, cf. output from summary_statistics)}

\item{post}{Dataframe of posterior parameter estimates (with columns Variable, Index, Mean and sd, , cf. output from summary_statistics)}

\item{param}{Vector of parameters' name to check}
}
\value{
Ggplot
}
\description{
Plot posterior shrinkage (capturing how much the model learns) vs Prior/Posterior "distance" (whether the prior "includes" the posterior).
See reference for the inspiration.
}
\references{
M. Betancourt, \href{https://betanalpha.github.io/assets/case_studies/principled_bayesian_workflow.html}{“Towards a Principled Bayesian Workflow”}, 2018.
}
