% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{compute_calibration}
\alias{compute_calibration}
\title{Estimate calibration given forecasts and corresponding outcomes}
\usage{
compute_calibration(forecast, outcome, res = NULL)
}
\arguments{
\item{forecast}{Vector of probability forecast}

\item{outcome}{Vector of observations (0 or 1)}

\item{res}{Resolution (bin's width). If NULL, automatic bin width selection with Sturges' method}
}
\value{
Dataframe with colums Forecast (bins), Frequency (frequency of outcomes in the bin), Lower (lower bound of the 95% CI) and Upper (upper bound of the 95% CI)
}
\description{
Estimate calibration given forecasts and corresponding outcomes
}
\examples{
N <- 10000
f <- runif(N)
o <- sapply(1:N, function(i) {rbinom(1, 1, f[i])})
cal <- compute_calibration(f, o)
with(cal, plot(Forecast, Frequency))
abline(c(0, 1), col = "red")
}
