% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_workflow.R
\name{compute_coverage}
\alias{compute_coverage}
\title{Compute coverage of CI for different confidence level}
\usage{
compute_coverage(post_samples, truth, CI = seq(0, 1, 0.05))
}
\arguments{
\item{post_samples}{Matrix of posterior samples. Rows represent a sample and columns represent variables.}

\item{truth}{Vector of true parameter values (should be the same length as the number of columns in post_samples).}

\item{CI}{Vector of confidence levels.}
}
\value{
Dataframe containing coverage (and uncertainty interval for the coverage) for different confidence level (nominal)
}
\description{
Useful for fake data check.
}
