% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_post.R
\name{compute_prior_influence}
\alias{compute_prior_influence}
\title{Compute diagnostics of how the posterior is influenced by the prior}
\usage{
compute_prior_influence(prior, post, pars)
}
\arguments{
\item{prior}{Dataframe of prior parameter estimates (with columns Variable, Index, Mean and sd, cf. output from summary_statistics)}

\item{post}{Dataframe of posterior parameter estimates (with columns Variable, Index, Mean and sd, , cf. output from summary_statistics)}

\item{pars}{Vector of parameters' name to check. Defaults to all parameters presents in post and prior.}
}
\value{
Dataframe with columns: Variable, Index, PostShrinkage, DistPrior
}
\description{
\itemize{
\item Posterior shrinkage (\code{PostShrinkage}) = 1 - Var(Post) / Var(Prior), capturing how much the model is learning.
Shrinkage near 0 indicates that the data provides little information beyond the prior.
Shrinkage near 1 indicates that the data is much more informative than the prior.
\item Mahalanobis distance between the mean posterior and the prior (\code{DistPrior}), capturing whether the prior "includes" the posterior.
}
}
\references{
M. Betancourt, \href{https://betanalpha.github.io/assets/case_studies/principled_bayesian_workflow.html}{“Towards a Principled Bayesian Workflow”}, 2018.
}
\seealso{
\code{\link[=plot_prior_influence]{plot_prior_influence()}} to directly plot PostShrink vs DistPrior
}
