% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{compute_RPS}
\alias{compute_RPS}
\title{Compute RPS for a single forecast}
\usage{
compute_RPS(forecast, outcome)
}
\arguments{
\item{forecast}{Vector of length N (forecast).}

\item{outcome}{Index of the true outcome (between 1 and N).}
}
\value{
RPS (numeric scalar)
}
\description{
Compute RPS for a single forecast
}
\examples{
compute_RPS(c(.2, .5, .3), 2)
}
