% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval.R
\name{empirical_pval}
\alias{empirical_pval}
\title{Compute empirical p-values}
\usage{
empirical_pval(t_rep, t, alternative = c("two.sided", "less", "greater"))
}
\arguments{
\item{t_rep}{Vector of samples from a distribution.}

\item{t}{Observation (numeric scalar).}

\item{alternative}{Indicates the alternative hypothesis: must be one of "two.sided", "greater" or "less".}
}
\value{
Empirical p-value.
}
\description{
Compute empirical p-values
}
\examples{
empirical_pval(rnorm(1e2), 2)
}
