% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_workflow.R
\name{extract_draws_from_array}
\alias{extract_draws_from_array}
\title{Extract parameters' draws from an array}
\usage{
extract_draws_from_array(obj, draws)
}
\arguments{
\item{obj}{Vector or matrix (columns represents different parameters) of draws}

\item{draws}{Vector draws to extract}
}
\value{
Dataframe with columns: Draw, Index, Value
}
\description{
Extract parameters' draws from an array
}
