% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{extract_index_nd}
\alias{extract_index_nd}
\title{Extract multiple indices inside bracket(s) as a list}
\usage{
extract_index_nd(x)
}
\arguments{
\item{x}{Character vector}
}
\value{
Dataframe with columns:
\itemize{
\item \code{Variable}, containing \code{x} where brackets have been removed
\item \code{Index}, a list containing values within the brackets
}
}
\description{
Extract multiple indices inside bracket(s) as a list
}
\examples{
extract_index_nd(c("sigma", "sigma[1]", "sigma[1, 1]", "sigma[1][2]"))
}
