% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_workflow.R
\name{extract_parameters_from_draw}
\alias{extract_parameters_from_draw}
\title{Extract parameters from a single draw}
\usage{
extract_parameters_from_draw(fit, param, draw)
}
\arguments{
\item{fit}{Stanfit object}

\item{param}{Vector of parameter names}

\item{draw}{Index of the draw to extract the parameters from}
}
\value{
Dataframe
}
\description{
Extract parameters from a single draw
}
\section{Note}{

Useful for to generate fake data.
}

