% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{extract_pdf}
\alias{extract_pdf}
\title{Extract probability density function from vector of samples}
\usage{
extract_pdf(x, support = NULL, n_density = 2^7)
}
\arguments{
\item{x}{Vector of samples from a distribution.}

\item{support}{Vector of length 2 corresponding to the range of the distribution. Can be NULL.}

\item{n_density}{Number of equally spaced points at which the density is to be estimated (better to use a power of 2).}
}
\value{
Dataframe with columns: \code{Value}, \code{Density.}
}
\description{
Extract probability density function from vector of samples
}
\examples{
extract_pdf(rnorm(1e4))
}
