% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illustrate_forward_chaining.R
\name{illustrate_forward_chaining}
\alias{illustrate_forward_chaining}
\title{Illustration forward chaining}
\usage{
illustrate_forward_chaining(horizon = 7, n_it = 5)
}
\arguments{
\item{horizon}{prediction horizon}

\item{n_it}{number of iterations to display}
}
\value{
Ggplot
}
\description{
Illustration forward chaining
}
\examples{

}
