% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_scalar}
\alias{is_scalar}
\title{Test whether x is of length 1}
\usage{
is_scalar(x)
}
\arguments{
\item{x}{Object to be tested.}
}
\value{
Logical
}
\description{
Test whether x is of length 1
}
\examples{
is_scalar(1) # TRUE
is_scalar("a") # TRUE
is_scalar(c(1, 2)) # FALSE
}
