% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_stanfit}
\alias{is_stanfit}
\title{Test whether an object is of class "stanfit"}
\usage{
is_stanfit(obj)
}
\arguments{
\item{obj}{Object.}
}
\value{
Boolean
}
\description{
Test whether an object is of class "stanfit"
}
