% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_wholenumber}
\alias{is_wholenumber}
\alias{is_scalar_wholenumber}
\title{Test whether x is a whole number}
\usage{
is_wholenumber(x, tol = .Machine$double.eps^0.5)

is_scalar_wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{Object to be tested}

\item{tol}{Tolerance}
}
\value{
Logical
}
\description{
Test whether x is a whole number
}
\examples{
is_wholenumber(1) # TRUE
is_wholenumber(1.0) # TRUE
is_wholenumber(1.1) # FALSE
is_scalar_wholenumber(1) # TRUE
is_scalar_wholenumber(c(1, 2)) # FALSE
}
