% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{logit}
\alias{logit}
\alias{inv_logit}
\title{Logit and Inverse logit}
\usage{
logit(x)

inv_logit(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric vector.
}
\description{
Logit and Inverse logit
}
\examples{
logit(0.5)
inv_logit(0)
}
