% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_workflow.R
\name{parameters_intervals}
\alias{parameters_intervals}
\title{Extract parameters distribution as a series of credible interval around the median.}
\usage{
parameters_intervals(fit, param, CI_width = seq(0.1, 0.9, 0.1))
}
\arguments{
\item{fit}{Stanfit object}

\item{param}{Vector of parameters to extract}

\item{CI_width}{Vector containing the width of the credible intervals}
}
\value{
(Long) Dataframe with columns: Variable, Index, Lower, Upper, Width
}
\description{
A potential improvement could be to compute the highest density posterior intervals (around the mode, narrowest but not symmetrical and harder to find).
}
