% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_workflow.R
\name{plot_coverage}
\alias{plot_coverage}
\title{Plot coverage of CI for different confidence level.
Useful for fake data check.}
\usage{
plot_coverage(post_samples, truth, CI = seq(0, 1, 0.05))
}
\arguments{
\item{post_samples}{Matrix of posterior samples. Rows represent a sample and columns represent variables.}

\item{truth}{Vector of true parameter values (should be the same length as the number of columns in post_samples).}

\item{CI}{Vector of confidence levels.}
}
\value{
Ggplot of coverage as a function of the nominal coverage (confidence level), with 95% uncertainty interval
}
\description{
Plot coverage of CI for different confidence level.
Useful for fake data check.
}
\examples{
N <- 100
N_post <- 1e3
truth <- rep(0, N)
post_samples <- sapply(rnorm(N, 0, 1), function(x) {rnorm(N_post, x, 1)})
plot_coverage(post_samples, truth)
}
