% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_post.R
\name{plot_prior_influence}
\alias{plot_prior_influence}
\alias{check_model_sensitivity}
\title{Plot prior influence diagnostics}
\usage{
plot_prior_influence(prior, post, param)

check_model_sensitivity(prior, post, param)
}
\arguments{
\item{prior}{Dataframe of prior parameter estimates (with columns Variable, Index, Mean and sd, cf. output from summary_statistics)}

\item{post}{Dataframe of posterior parameter estimates (with columns Variable, Index, Mean and sd, , cf. output from summary_statistics)}

\item{param}{Vector of parameters' name to check}
}
\value{
Ggplot
}
\description{
Plot posterior shrinkage (capturing how much the model learns) vs Mahalanobis distance between the mean posterior and the prior (whether the prior "includes" the posterior).
}
\seealso{
\code{\link[=compute_prior_influence]{compute_prior_influence()}} to return a dataframe with posterior shrinkage and prior/posterior distance.
}
