% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_post.R
\name{plot_prior_posterior}
\alias{plot_prior_posterior}
\title{Plot posterior CI alongside prior CI}
\usage{
plot_prior_posterior(prior, post, pars = NULL)
}
\arguments{
\item{prior}{Dataframe of prior parameter estimates}

\item{post}{Dataframe of posterior parameter estimates}

\item{pars}{Vector of parameter names to plot. Defaults to all parameters presents in post and prior.}
}
\value{
Ggplot of parameter estimates
}
\description{
Plot posterior CI alongside prior CI
}
