% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_post.R
\name{prior_posterior}
\alias{prior_posterior}
\alias{combine_prior_posterior}
\alias{plot_prior_posterior}
\alias{compute_prior_influence}
\alias{plot_prior_influence}
\alias{check_model_sensitivity}
\title{Compare prior to posterior}
\usage{
combine_prior_posterior(prior, post, pars = NULL, match_exact = TRUE)

plot_prior_posterior(prior, post, pars = NULL, match_exact = TRUE)

compute_prior_influence(prior, post, pars = NULL, match_exact = TRUE)

plot_prior_influence(prior, post, pars = NULL, match_exact = TRUE)

check_model_sensitivity(prior, post, pars = NULL)
}
\arguments{
\item{prior}{Dataframe of prior parameter estimates}

\item{post}{Dataframe of posterior parameter estimates}

\item{pars}{Vector of parameter names to plot. Defaults to all parameters presents in \code{post} and \code{prior.}}

\item{match_exact}{Logical indicating whether parameters should be matched exactly (e.g. "p" does not match "p[1]").}
}
\value{
\itemize{
\item \code{combine_prior_posterior} returns a dataframe with the same columns as in prior and post + "Distribution".
\item \code{compute_prior_influence} returns a dataframe with columns: Variable, Index, PostShrinkage, DistPrior.
\item \code{plot_prior_posterior} and \code{plot_prior_influence} returns a ggplot object
}
}
\description{
\itemize{
\item \code{combine_prior_posterior} subsets and binds the prior and posterior dataframes
\item \code{plot_prior_posterior} plots posterior CI alongside prior CI
\item \code{compute_prior_influence} computes diagnostics of how the posterior is influenced by the prior
\item \code{plot_prior_influence} plots diagnostics from \code{compute_prior_influence}
}
}
\details{
\itemize{
\item Posterior shrinkage (\code{PostShrinkage}) = 1 - Var(Post) / Var(Prior), capturing how much the model is learning.
Shrinkage near 0 indicates that the data provides little information beyond the prior.
Shrinkage near 1 indicates that the data is much more informative than the prior.
\item Mahalanobis distance between the mean posterior and the prior (\code{DistPrior}), capturing whether the prior "includes" the posterior.
}
}
\references{
M. Betancourt,  \href{https://betanalpha.github.io/assets/case_studies/principled_bayesian_workflow.html}{“Towards a Principled Bayesian Workflow”}, 2018.
}
