% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{change_colnames}
\alias{change_colnames}
\title{Change column names of a dataframe}
\usage{
change_colnames(df, current_names, new_names)
}
\arguments{
\item{df}{Dataframe}

\item{current_names}{Vector of column names to change}

\item{new_names}{Vector of new names}
}
\value{
Dataframe with new column names
}
\description{
Change column names of a dataframe
}
\examples{
df <- data.frame(A = 1:2, B = 3:4, C = 5:6)
df <- change_colnames(df, c("A", "C"), c("Aa", "Cc"))
}
