% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_workflow.R
\name{extract_distribution}
\alias{extract_distribution}
\title{Extract a distribution represented by samples}
\usage{
extract_distribution(
  object,
  parName,
  type = c("continuous", "discrete", "samples", "eti", "hdi"),
  support = NULL,
  transform = identity,
  nDensity = 2^7,
  nDraws = 100,
  CI_level = seq(0.1, 0.9, 0.1)
)
}
\arguments{
\item{object}{Object specifying the distribution as samples: can be a Stanfit object, a matrix (columns represents parameters, rows samples) or a vector.}

\item{parName}{Name of the parameter to extract}

\item{type}{Indicates how the distribution is summarised.}

\item{support}{Support of the distribution. For type = "continuous", this must be the range of the distribution. For type = "discrete", this must be a vector of all possible values that the distribution can take. Can be NULL.}

\item{transform}{Function to apply to the samples}

\item{nDensity}{Number of equally spaced points at which the density is to be estimated (better to use a power of 2). Applies when type = "continuous".}

\item{nDraws}{Number of draws from the distribution. Applies when type = "samples"}

\item{CI_level}{Vector containing the level of the confidence/credible intervals}
}
\value{
Dataframe
}
\description{
The distribution can be extracted as:
\itemize{
\item a probability density function ("continuous").
\item a probability mass unction ("discrete").
\item draws from the distribution ("samples"), in that case, little processing is done.
\item a series of equal-tailed confidence/credible intervals ("eti").
\item a series of highest density confidence/credible intervals ("hdi").
}
}
