% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_workflow.R
\name{extract_draws}
\alias{extract_draws}
\title{Extract parameters' draws}
\usage{
extract_draws(obj, draws)
}
\arguments{
\item{obj}{Vector, matrix (columns represents different parameters) of draws or list of it}

\item{draws}{Vector of draws to extract}
}
\value{
Dataframe with columns: Draw, Index, Value, Parameter
}
\description{
Extract parameters' draws
}
\examples{
x <- rnorm(1e3)
X <- matrix(x, ncol = 10)
extract_draws(x, sample(1:length(x), 10))
extract_draws(X, sample(1:nrow(X), 10))
extract_draws(list(x = x, X = X), 1:10)
}
