% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{extract_index_nd}
\alias{extract_index_nd}
\title{Extract multiple indices inside bracket(s) as a list}
\usage{
extract_index_nd(x)
}
\arguments{
\item{x}{Character vector}
}
\value{
Dataframe with columns Variable and Index
}
\description{
Extract multiple indices inside bracket(s) as a list
}
\examples{
extract_index_nd(c("sigma", "sigma[1]", "sigma[1, 1]", "sigma[1][2]"))
}
