% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{factor_to_numeric}
\alias{factor_to_numeric}
\title{Change the type of the column of a dataframe from factor to numeric}
\usage{
factor_to_numeric(df, factor_name)
}
\arguments{
\item{df}{Dataframe}

\item{factor_name}{Name of the factor to change into a numeric}
}
\value{
Same dataframe with type of the given column changed to numeric
}
\description{
Change the type of the column of a dataframe from factor to numeric
}
\examples{
df <- data.frame(A = rep(1:5, each = 10))
df$A <- factor(df$A)
df <- factor_to_numeric(df, "A")
}
