% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{factor_to_numeric}
\alias{factor_to_numeric}
\title{Change the type of the column of a dataframe from factor to numeric}
\usage{
factor_to_numeric(df, factor_name)
}
\arguments{
\item{df}{Dataframe.}

\item{factor_name}{Vector of names of factors to change to numeric.}
}
\value{
Same dataframe with type of the given columns changed to numeric.
}
\description{
Change the type of the column of a dataframe from factor to numeric
}
\examples{
df <- data.frame(A = rep(1:5, each = 10))
df$A <- factor(df$A)
df <- factor_to_numeric(df, "A")
}
