% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HuraultMisc-package.R
\docType{package}
\name{HuraultMisc-package}
\alias{HuraultMisc}
\alias{HuraultMisc-package}
\title{HuraultMisc: Guillem Hurault Personal Functions' Library}
\description{
Contains various functions for data analysis, notably helpers and diagnostics for Bayesian modelling using Stan.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ghurault/HuraultMisc}
  \item Report bugs at \url{https://github.com/ghurault/HuraultMisc/issues}
}

}
\author{
\strong{Maintainer}: Guillem Hurault \email{guillem.hurault@hotmail.fr} (\href{https://orcid.org/0000-0002-1052-3564}{ORCID})

}
\keyword{internal}
