% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illustrations.R
\name{illustrate_forward_chaining}
\alias{illustrate_forward_chaining}
\title{Illustration forward chaining}
\usage{
illustrate_forward_chaining(horizon = 7, n_it = 5)
}
\arguments{
\item{horizon}{Prediction horizon}

\item{n_it}{Number of iterations to display}
}
\value{
Ggplot
}
\description{
Illustration forward chaining
}
\examples{
illustrate_forward_chaining()

}
