% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illustrations.R
\name{illustrate_RPS}
\alias{illustrate_RPS}
\title{Illustration of the Ranked Probability Score}
\usage{
illustrate_RPS(mu = 5, sigma = 1, observed = 6)
}
\arguments{
\item{mu}{Mean of the Gaussian forecast distribution}

\item{sigma}{Standard deviation of the Gaussian forecast distribution}

\item{observed}{Observed outcome}
}
\value{
ggplot
}
\description{
Illustrates what the RPS measure when we forecast "Severity" with a Gaussian distribution.
The RPS corresponds to the shaded area squared.
}
\examples{
illustrate_RPS()

}
