% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_workflow.R
\name{parameters_intervals}
\alias{parameters_intervals}
\title{Extract parameters' credible interval}
\usage{
parameters_intervals(
  fit,
  param,
  CI_width = seq(0.1, 0.9, 0.1),
  type = c("hdi", "eti")
)
}
\arguments{
\item{fit}{Stanfit object}

\item{param}{Vector of parameters to extract}

\item{CI_width}{Vector containing the width of the credible intervals}

\item{type}{Should be one of hdi (highest density interval, i.e. narrowest) or eti (equal-tailed interval, i.e. centered around median)}
}
\value{
(Long) Dataframe with columns: Variable, Index, Lower, Upper, Width
}
\description{
Extract parameters distribution as a series of credible interval (equal-tailed or highest density)
}
\section{Note}{

This function should not work for more than one-dimensional parameters
}

