% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_workflow.R
\name{plot_prior_posterior}
\alias{plot_prior_posterior}
\title{Plot posterior estimates alongside prior estimates}
\usage{
plot_prior_posterior(post, prior, param)
}
\arguments{
\item{post}{Dataframe of posterior parameter estimates}

\item{prior}{Dataframe of prior parameter estimates}

\item{param}{Vector of parameter names to plot}
}
\value{
ggplot of parameter estimates
}
\description{
Plot posterior estimates alongside prior estimates
}
