% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_functions.R
\name{PPC_group_distribution}
\alias{PPC_group_distribution}
\title{Posterior Predictive Check for Stan model}
\usage{
PPC_group_distribution(fit, parName, nDraws = 1)
}
\arguments{
\item{fit}{Stanfit object}

\item{parName}{Name of the observation-dependent (e.g. patient-dependent) parameter to consider}

\item{nDraws}{Number of draws to plot}
}
\value{
ggplot of the distribution
}
\description{
Plot the distribution density of parameters within a same group from a single/multiple draw of the posterior distribution.
In the case of a hierarchical model, we might look at the distribution of patient parameter and compare it to the prior for the population distribution.
}
