% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPC_group_distribution.R
\name{PPC_group_distribution}
\alias{PPC_group_distribution}
\title{Plot the distribution density of parameters within a same group from a single/multiple draw of the posterior distribution.
In the case of a hierarchical model, we might look at the distribution of patient parameter and compare it to the prior for the population distribution.}
\usage{
PPC_group_distribution(fit, parName, nDraws = 1)
}
\arguments{
\item{fit}{stanfit object}

\item{parName}{name of the observation-dependent (e.g. patient-dependent) parameter to consider}

\item{nDraws}{number of draws to plot}
}
\value{
Ggplot of the distribution
}
\description{
Plot the distribution density of parameters within a same group from a single/multiple draw of the posterior distribution.
In the case of a hierarchical model, we might look at the distribution of patient parameter and compare it to the prior for the population distribution.
}
\examples{

}
