% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_workflow.R
\name{process_replications}
\alias{process_replications}
\title{Extract posterior predictive distribution}
\usage{
process_replications(
  fit,
  idx = NULL,
  parName,
  type = c("continuous", "discrete", "samples", "eti", "hdi"),
  bounds = NULL,
  nDensity = 2^7,
  nDraws = 100,
  CI_level = seq(0.1, 0.9, 0.1)
)
}
\arguments{
\item{fit}{Stanfit object}

\item{idx}{Dataframe for translating the indices of the replication parameters into more informative variable (can be NULL)}

\item{parName}{Name of the replication parameter}

\item{type}{Indicates how the distribution is summarised. Values can take "continuous", "discrete" or "samples".}

\item{bounds}{NULL or vector of length 2 representing the bounds of the distribution if it needs to be truncated.}

\item{nDensity}{Number of equally spaced points at which the density is to be estimated (better to use a power of 2). Applies when type = "continuous".}

\item{nDraws}{Number of draws from the distribution. Applies when type = "samples"}

\item{CI_level}{Vector containing the level of the confidence/credible intervals}
}
\value{
Dataframe
}
\description{
The Posterior predictive distribution is extracted as a "continuous" function (cf. density), as a "discrete" function (cf. probability) or as "samples" (draws from the distribution)
}
