% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_functions.R
\name{process_replications}
\alias{process_replications}
\title{Extract posterior predictive distribution}
\usage{
process_replications(fit, idx, parName, type = "continuous", bounds,
  nDensity = 2^7, nDraws = 100)
}
\arguments{
\item{fit}{Stanfit object}

\item{idx}{Dataframe for translating the indices of the replication parameters into more informative variable}

\item{parName}{Name of the replication parameter}

\item{type}{Indicates how the distribution is summarised. Values can take "continuous", "discrete" or "samples".}

\item{bounds}{Bounds of the distribution (vector of length 2)}

\item{nDensity}{Number of equally spaced points at which the density is to be estimated (better to use a power of 2). Applies when type = "continuous".}

\item{nDraws}{Number of draws from the distribution. Applies when type = "samples"}
}
\value{
Dataframe
}
\description{
The Posterior predictive distribution is extracted as a "continuous" function (cf. density), as a "discrete" function (cf. probability) or as "samples" (draws from the distribution)
}
